iatp-mock/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.111.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.4
orjson==3.9.15
pycparser==2.22
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.7
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.37.2
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.27.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
